- move validations to utils folder
